+++ libraries/Firebase_ESP32_Client/src/wcs/esp32/FB_TCP_Client.h
@@ -1,10 +1,10 @@
 /**
- * Firebase TCP Client v1.1.22
+ * Firebase TCP Client v1.1.23
  *
- * Created July 10, 2022
+ * Created December 19, 2022
  *
  * The MIT License (MIT)
- * Copyright (c) 2022 K. Suwatchai (Mobizt)
+ * Copyright (c) 2023 K. Suwatchai (Mobizt)
  *
  *
  * Copyright (c) 2015 Markus Sattler. All rights reserved.